+++ util/options.cc
@@ -20,8 +20,8 @@
 #include "rocksdb/cache.h"
 #include "rocksdb/compaction_filter.h"
 #include "rocksdb/comparator.h"
-#include "rocksdb/delete_scheduler.h"
 #include "rocksdb/env.h"
+#include "rocksdb/sst_file_manager.h"
 #include "rocksdb/memtablerep.h"
 #include "rocksdb/merge_operator.h"
 #include "rocksdb/slice.h"
@@ -213,7 +213,7 @@ DBOptions::DBOptions()
       paranoid_checks(true),
       env(Env::Default()),
       rate_limiter(nullptr),
-      delete_scheduler(nullptr),
+      sst_file_manager(nullptr),
       info_log(nullptr),
 #ifdef NDEBUG
       info_log_level(INFO_LEVEL),
@@ -281,7 +281,7 @@ DBOptions::DBOptions(const Options& options)
       paranoid_checks(options.paranoid_checks),
       env(options.env),
       rate_limiter(options.rate_limiter),
-      delete_scheduler(options.delete_scheduler),
+      sst_file_manager(options.sst_file_manager),
       info_log(options.info_log),
       info_log_level(options.info_log_level),
       max_open_files(options.max_open_files),
@@ -433,8 +433,9 @@ void DBOptions::Dump(Logger* log) const {
         use_adaptive_mutex);
     Header(log, "                            Options.rate_limiter: %p",
         rate_limiter.get());
-    Header(log, "     Options.delete_scheduler.rate_bytes_per_sec: %" PRIi64,
-         delete_scheduler ? delete_scheduler->GetRateBytesPerSecond() : 0);
+    Header(
+        log, "     Options.sst_file_manager.rate_bytes_per_sec: %" PRIi64,
+        sst_file_manager ? sst_file_manager->GetDeleteRateBytesPerSecond() : 0);
     Header(log, "                          Options.bytes_per_sync: %" PRIu64,
         bytes_per_sync);
     Header(log, "                      Options.wal_bytes_per_sync: %" PRIu64,